openapi importer properly detects json request and response
